remove redundant vc
