+++ NewsApp/NewsApp/Dashboard/DashboardViewController.swift
@@ -9,8 +9,27 @@ import Foundation
 import UIKit
 
 class DashboardViewController: UIViewController {
-    let discoverViewController = DiscoverNewsViewController()
-    let feedViewController = NewsFeedViewController()
+    private let articleDataSource = NewsArticleDataSource()
+    private let feedTableView = NewsFeedTableView()
+    private let newsSourcesViewController = NewsSourcesViewController()
+    
+    private lazy var refreshControl: UIRefreshControl = {
+        let refreshControl = UIRefreshControl()
+        refreshControl.addTarget(self, action: #selector(syncTable), for: UIControl.Event.valueChanged)
+        return refreshControl
+    }()
+    
+    init() {
+        super.init(nibName: nil, bundle: nil)
+        
+        articleDataSource.delegate = self
+        articleDataSource.syncArticles(forCountry: .us)
+        articleDataSource.loadArticles()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -24,21 +43,28 @@ class DashboardViewController: UIViewController {
         let sideMenuButtonBarButtonItem = UIBarButtonItem(customView: sideMenuButton)
         navigationItem.rightBarButtonItem = sideMenuButtonBarButtonItem
         navigationItem.title = "Dashboard"
+
+        feedTableView.dataSource = self
+        feedTableView.delegate = self
+        feedTableView.addSubview(refreshControl)
         
-        let newsFeedContentView = UIView()
-        newsFeedContentView.translatesAutoresizingMaskIntoConstraints = false
+        view.addSubview(feedTableView)
         
-        view.addSubview(newsFeedContentView)
-        
-        add(asChildViewController: feedViewController, contentView: newsFeedContentView)
-        
-        let safeArea = view.safeAreaLayoutGuide
         NSLayoutConstraint.activate([
-            newsFeedContentView.topAnchor.constraint(equalTo: safeArea.topAnchor),
-            newsFeedContentView.leadingAnchor.constraint(equalTo: safeArea.leadingAnchor),
-            newsFeedContentView.trailingAnchor.constraint(equalTo: safeArea.trailingAnchor),
-            newsFeedContentView.bottomAnchor.constraint(equalTo: safeArea.bottomAnchor),
+            feedTableView.topAnchor.constraint(equalTo: view.topAnchor),
+            feedTableView.leftAnchor.constraint(equalTo: view.leftAnchor),
+            feedTableView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            feedTableView.rightAnchor.constraint(equalTo: view.rightAnchor),
         ])
+    }
+    
+    @objc private func syncTable() {
+        // TODO: what do we sync and load exactly ?
+        articleDataSource.syncArticles(forCountry: .de, completion: { [weak self] in
+            DispatchQueue.main.async {
+                self?.refreshControl.endRefreshing()
+            }
+        })
     }
     
     @objc private func showRightSideBar() {
@@ -46,3 +72,62 @@ class DashboardViewController: UIViewController {
         sideMenuController.isMenuRevealed ? sideMenuController.hideMenu() : sideMenuController.revealMenu()
     }
 }
+
+extension DashboardViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return articleDataSource.articles?.count ?? 0
+    }
+
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: NewsFeedTableViewCell.reuseIdentifier, for: indexPath) as? NewsFeedTableViewCell else {
+            return UITableViewCell()
+        }
+        
+        cell.setupCell(article: articleDataSource.articles?[indexPath.row] ?? nil)
+        
+        return cell
+    }
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        guard let article = articleDataSource.articles?[indexPath.row] else { return }
+        
+        NewsAPISyncer().getImage(forUrl: article.urlToImage, completion: { [weak self] image in
+            let newsDetailsViewController = NewsDetailsViewController(withArticle: article, posterImage: image)
+            
+            DispatchQueue.main.async {
+                self?.navigationController?.pushViewController(newsDetailsViewController, animated: true)
+                tableView.cellForRow(at: indexPath)?.isSelected = false
+            }
+        })
+        
+    }
+}
+
+extension DashboardViewController: UITableViewDelegate {
+    func numberOfSections(in tableView: UITableView) -> Int {
+        1
+    }
+    
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        150
+    }
+    
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        let contentView = UIView()
+        
+        add(asChildViewController: newsSourcesViewController, contentView: contentView)
+        
+        return contentView
+    }
+}
+
+extension DashboardViewController: NewsArticleDataSourceDelegate {
+    func newsArticleDataSourceDeletage(willUpdateArticles dataSource: NewsArticleDataSource) {}
+    
+    func newsArticleDataSourceDelegate(didUpdateArticles dataSource: NewsArticleDataSource) {
+        DispatchQueue.main.async { [weak self] in
+            self?.feedTableView.reloadData()
+        }
+    }
+}
+
